fix: extract kanban session diagnostics
